+++ src/main/java/com/programmers/voucher/domain/voucher/dto/request/VoucherCreateRequest.java
@@ -1,7 +1,7 @@
 package com.programmers.voucher.domain.voucher.dto.request;
 
 import com.programmers.voucher.domain.voucher.domain.VoucherType;
-import com.programmers.voucher.global.validation.VoucherTypeAmountValidation;
+import com.programmers.voucher.domain.voucher.dto.request.validation.VoucherTypeAmountValidation;
 import jakarta.validation.constraints.NotNull;
 
 @VoucherTypeAmountValidation
